PNP implemented
The parameters for ransac are chosen arbitrarly, but they gave kind of good results for KITTI. Important: the rotation_vector that gets returned from opencv is actually mapping from world to camj. Also, I remove at the landmarks and keypoints based on the results that the ransac gave me. This ensures that the number of landmark to kps correspondences reduces. Works for all datasets except woko. There the success variable suddenly gets false. Not sure why.